[Refactor] : ADViewCase 추가
Git : #66
