+++ Projects/Domain/Domain_Model/Sources/Entity/ADViewCase.swift
@@ -1,0 +1,15 @@
+//
+//  ADViewCase.swift
+//  Domain_Model
+//
+//  Created by minii on 2023/10/09.
+//  Copyright © 2023 chminipark. All rights reserved.
+//
+
+import Foundation
+
+public enum ADViewCase: Equatable {
+  case OnBoarding
+  case MakeAD
+  case ConfigureAnimation
+}
