+++ Projects/Domain/SharedProvider/Sources/SharedProvider.swift
@@ -15,6 +15,7 @@ import ComposableArchitecture
 public struct Shared {
   public let makeAD = Self.MakeAD()
   public let stepBar = Self.StepBar()
+  public let adViewCase = CombineNotifier<ADViewCase>(initialValue: .OnBoarding)
 }
 
 public extension Shared {
@@ -32,7 +33,6 @@ public extension Shared {
     public let croppedImage = CombineNotifier<UIImage?>(initialValue: nil)
     public let maskedImage = CombineNotifier<UIImage?>(initialValue: nil)
     public let jointsDTO = CombineNotifier<JointsDTO?>(initialValue: nil)
-    public let isShowConfigureAnimationView = CombineNotifier<Bool>(initialValue: false)
   }
 }
