+++ Projects/Presentation/ConfigureAnimation/Feature/ConfigureAnimationFeature.swift
@@ -18,6 +18,7 @@ public struct ConfigureAnimationFeature: Reducer {
   @Dependency(\.configureAnimationClient) var configureAnimationClient
   @Dependency(\.shared.stepBar) var stepBar
   @Dependency(\.shared.makeAD) var makeAD
+  @Dependency(\.shared.adViewCase) var adViewCase
   
   public init() {}
   
@@ -87,7 +88,7 @@ extension ConfigureAnimationFeature {
         
       case .fixMakeAD:
         return .run { _ in
-          await makeAD.isShowConfigureAnimationView.set(false)
+          await adViewCase.set(.MakeAD)
         }
         
       case .toggleIsShowAnimationListView:
@@ -252,7 +253,7 @@ extension ConfigureAnimationFeature {
           await makeAD.croppedImage.set(nil)
           await makeAD.maskedImage.set(nil)
           await makeAD.jointsDTO.set(nil)
-          await makeAD.isShowConfigureAnimationView.set(false)
+          await adViewCase.set(.MakeAD)
           
           await stepBar.completeStep.set(.None)
           await stepBar.currentStep.set(.UploadADrawing)
