+++ Projects/Presentation/FindingTheCharacter/Feature/FindingTheCharacterFeature.swift
@@ -50,9 +50,6 @@ public struct FindingTheCharacterFeature: Reducer {
     
     case alertShared(PresentationAction<AlertShared>)
     case showAlertShared(AlertState<AlertShared>)
-    
-    case initCheckList
-    case isCorrectStep
   }
   
   public var body: some Reducer<State, Action> {
@@ -167,20 +164,6 @@ extension FindingTheCharacterFeature {
       case .showAlertShared(let alertState):
         state.alertShared = alertState
         return .none
-        
-        
-      case .initCheckList:
-        state.checkState = false
-        return .none
-        
-      case .isCorrectStep:
-        return .run { send in
-          let completeStep = await stepBar.completeStep.get()
-          let isCorrectStep = Step.isCorrectStep(myStep: .UploadADrawing, completeStep: completeStep)
-          if !isCorrectStep {
-            await send(.initCheckList)
-          }
-        }
       }
     }
   }
