+++ Projects/Presentation/RootView/View/RootView.swift
@@ -10,40 +10,43 @@ import SwiftUI
 import OnBoarding
 import MakeAD
 import ConfigureAnimation
-import ThirdPartyLib
 import SharedProvider
-import RootView_Feature
+import Domain_Model
 
-public struct RootView: ADUI {
-  public typealias MyFeature = RootViewFeature
-  @StateObject var viewStore: MyViewStore
+let onBoardingViewController: UIHostingController<OnBoardingView> = .init(
+  rootView: OnBoardingView()
+)
+let makeADViewController: UIHostingController<MakeADView> = .init(
+  rootView: MakeADView()
+)
+let configureAnimationViewController: UIHostingController<ConfigureAnimationView> = .init(
+  rootView: ConfigureAnimationView()
+)
+
+public struct RootView: View {
+  public init() {}
   
-  public init(
-    viewStore: MyViewStore = .init(
-      Store(initialState: .init(), reducer: {
-        MyFeature()
-      }),
-      observe: { $0 }
-    )
-  ) {
-    self._viewStore = StateObject(wrappedValue: viewStore)
-  }
-  
-  @SharedValue(\.shared.makeAD.isShowConfigureAnimationView) var isShowConfigureAnimationView
-  
-  var switchValue: ADViewCase {
-    isShowConfigureAnimationView ? .ConfigureAnimation : .MakeAD
-  }
+  @SharedValue(\.shared.adViewCase) var adViewCase
   
   public var body: some View {
-    Group {
-      if !(viewStore.isTapStartButton) {
-        OnBoardingView(isTapStartButton: viewStore.$isTapStartButton)
-      } else {
-        SwitchView(switchValue: switchValue)
-          .ignoresSafeArea()
-      }
+    ZStack {
+      SwitchHostingView(
+        vc: onBoardingViewController,
+        switchValue: adViewCase,
+        mySwitchValue: ADViewCase.OnBoarding
+      )
+      SwitchHostingView(
+        vc: makeADViewController,
+        switchValue: adViewCase,
+        mySwitchValue: ADViewCase.MakeAD
+      )
+      SwitchHostingView(
+        vc: configureAnimationViewController,
+        switchValue: adViewCase,
+        mySwitchValue: ADViewCase.ConfigureAnimation
+      )
     }
+    .ignoresSafeArea()
   }
 }
